🕸️ Weaver: Refactor ScheduledTaskModal using Custom Hook Pattern
