+++ src/features/scheduled-tasks/components/ScheduledTaskModal.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState, type RefObject } from 'react';
+import { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { open } from '@tauri-apps/plugin-dialog';
 import { toast } from 'sonner';
@@ -27,11 +27,7 @@ import { MentionTextarea } from './MentionTextarea';
 import { getDisplayCron, ScheduleBuilder } from './ScheduleBuilder';
 import type { Assistant } from '@/models/chat';
 import { getLogger } from '@/lib/logger';
-import { useDnDContext } from '@/context/DnDContext';
-import {
-  checkDroppedPathType,
-  registerDroppedFiles,
-} from '@/lib/backend/file-operations';
+import { useWorkspaceDropZone } from '../hooks/useWorkspaceDropZone';
 
 const logger = getLogger('ScheduledTaskModal');
 
@@ -111,7 +107,6 @@ function ScheduledTaskForm({
   onSave,
 }: ScheduledTaskFormProps) {
   const { t } = useTranslation();
-  const { subscribe } = useDnDContext();
   const [name, setName] = useState(task?.name ?? '');
   const initialCronExpression =
     task !== undefined && task !== null
@@ -145,72 +140,14 @@ function ScheduledTaskForm({
   const [workspaceOverride, setWorkspaceOverride] = useState<string | null>(
     task?.workspaceOverride ?? null,
   );
-  const [workspaceDragState, setWorkspaceDragState] = useState<
-    'none' | 'valid' | 'invalid'
-  >('none');
   const [browsingWorkspace, setBrowsingWorkspace] = useState(false);
   const [saving, setSaving] = useState(false);
   const workspaceDropRef = useRef<HTMLDivElement>(null);
 
-  useEffect(() => {
-    const processDroppedPaths = (paths: string[]) => {
-      const run = async () => {
-        try {
-          await registerDroppedFiles(paths);
-        } catch (error: unknown) {
-          logger.error('Failed to register dropped workspace folder', error);
-          toast.error(t('scheduledTasks.modal.workspaceRegisterFailed'));
-          return;
-        }
-
-        for (const filePath of paths) {
-          try {
-            const pathType = await checkDroppedPathType(filePath);
-            if (pathType === 'directory') {
-              setWorkspaceOverride(filePath);
-              return;
-            }
-          } catch (error: unknown) {
-            logger.error('Failed to inspect dropped workspace path', {
-              filePath,
-              error,
-            });
-          }
-        }
-
-        toast.error(t('scheduledTasks.modal.workspaceDropFolderError'));
-      };
-
-      void run();
-    };
-
-    const unsubscribe = subscribe(
-      workspaceDropRef as RefObject<HTMLElement>,
-      (event, payload) => {
-        if (event === 'drag-over') {
-          setWorkspaceDragState(
-            payload.paths && payload.paths.length > 0 ? 'valid' : 'invalid',
-          );
-          return;
-        }
-
-        if (event === 'leave') {
-          setWorkspaceDragState('none');
-          return;
-        }
-
-        setWorkspaceDragState('none');
-        if (payload.paths && payload.paths.length > 0) {
-          processDroppedPaths(payload.paths);
-        }
-      },
-      { priority: 5 },
-    );
-
-    return () => {
-      unsubscribe();
-    };
-  }, [subscribe, t]);
+  const { workspaceDragState } = useWorkspaceDropZone(
+    workspaceDropRef,
+    setWorkspaceOverride,
+  );
 
   const handleBrowseWorkspace = async () => {
     if (browsingWorkspace) return;
